Factorize UserRouter into EventHandlersRouter
