+++ fastapi_users/router/__init__.py
@@ -1,2 +1,6 @@
-from fastapi_users.router.common import ErrorCode, Event  # noqa: F401
-from fastapi_users.router.users import UserRouter, get_user_router  # noqa: F401
+from fastapi_users.router.common import (  # noqa: F401
+    ErrorCode,
+    Event,
+    EventHandlersRouter,
+)
+from fastapi_users.router.users import get_user_router  # noqa: F401
